refactor: My 페이지 관련 기존 코드 삭제 후 리팩토링 코드로 수정
