+++ src/App.js
@@ -19,9 +19,7 @@ function App() {
           <Route path="post/:postId" element={<MyDetailPage />}></Route>
         </Route>
         <Route path="/profile/setting" element={<SettingPage />}></Route>
-        <Route path="/test" element={<TestPage />}>
-          <Route path="post/:postId" element={<TestPage />}></Route>
-        </Route>
+        <Route path="/test" element={<TestPage />}></Route>
       </Routes>
       <NavigationBar />
     </BrowserRouter>
