+++ src/pages/MyPage/MyPage.jsx
@@ -1,45 +1,103 @@
 import React from "react";
 import Layout from "../../components/Layout";
-import Container from "./components/Container";
-import AccountInfoTemplate from "./components/AccountInfoTemplate";
-import MyAlbumTemplate from "./components/MyAlbumTemplate";
+import styled from "styled-components";
+import UploadButton from "./components/UploadButton";
+import KaKaoProfile from "./components/KaKaoProfile";
+import InstaProfile from "./components/InstaProfile";
+import Card from "./components/Card";
 import { AnimatePresence, motion } from "framer-motion";
-import { styled } from "styled-components";
-import MyDetailPage from "../MyDetailpage/MyDetailPage";
 import { useMatch, useNavigate } from "react-router-dom";
+import MyDetailPage from "../MyDetailpage/MyDetailPage";
 
-const userInfo = {
-  profileImg: "./icons/accountIcon.png",
-  userName: "카카오",
-  userEmail: "kakao@email.com",
+const response = {
+  id: 14,
+  username: "춘식이",
+  email: "chooonsik@naver.com",
+  profileImage: "sample.png",
+  instagram: {
+    isLinked: true,
+    infos: {
+      totalLikes: 54,
+      totalViews: 48,
+      fireworks: 700,
+    },
+    photos: [
+      {
+        id: 4,
+        image: "/sample.png",
+        createdAt: "2023-09-02",
+        name: "안녕하세요",
+        hashtags: ["해쉬태그", "안녕"],
+        likes: 50,
+        instaViews: 39,
+        isLiked: true,
+      },
+      {
+        id: 19,
+        image: "/sample3.png",
+        createdAt: "2023-10-17",
+        name: "열두글자이름이름열두글자",
+        hashtags: ["여덟글자해쉬태그", "안녕안녕"],
+        likes: 4,
+        instaViews: 0,
+        isLiked: true,
+      },
+      {
+        id: 21,
+        image: "/sample4.png",
+        createdAt: "2023-11-02",
+        name: "english Name",
+        hashtags: ["english"],
+        likes: 0,
+        instaViews: 9,
+        isLiked: false,
+      },
+    ],
+  },
 };
 
-const totalCount = {
-  totalLikes: 60,
-  totalFirework: 30,
-  myFireworks: 200,
+const response2 = {
+  id: 14,
+  username: "춘식이",
+  email: "chooonsik@naver.com",
+  profileImage: "/icons/accountIcon.png",
+  instagram: {
+    isLinked: false,
+    infos: {},
+    photos: [],
+  },
 };
 
-const isInstaConnected = true;
+const response3 = {
+  id: 14,
+  username: "춘식이",
+  email: "chooonsik@naver.com",
+  profileImage: "/icons/accountIcon.png",
+  instagram: {
+    isLinked: true,
+    infos: {
+      totalLikes: 0,
+      totalViews: 0,
+      fireworks: 300,
+    },
+    photos: [],
+  },
+};
 
-const photos = [
-  {
-    id: 1,
-    imgSrc: "/sample.png",
-    univercity: "전남대학교",
-    date: "2023.08.23",
-    likeCount: "60",
-    snsViewCount: "17",
-  },
-  {
-    id: 2,
-    imgSrc: "/sample5.png",
-    univercity: "전남대학교",
-    date: "2023.08.22",
-    likeCount: "452",
-    snsViewCount: "124",
-  },
-];
+const Container = styled.div`
+  width: 358px;
+  position: absolute;
+  top: 16px;
+  left: 50%;
+  transform: translateX(-50%);
+`;
+
+const Album = styled.div`
+  margin-top: 8px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+`;
 
 const Overlay = styled(motion.div)`
   position: fixed;
@@ -61,29 +119,53 @@ const Detail = styled(motion.div)`
   height: 80dvh;
   border-radius: 8px;
   background-color: rgb(15, 15, 15);
+  overflow: hidden;
 `;
 
 const MyPage = () => {
+  const {
+    username,
+    email,
+    profileImage,
+    instagram: { isLinked, infos, photos },
+  } = response;
   const navigate = useNavigate();
   const detailMatch = useMatch("/profile/post/:postId");
 
-  const clickedPhoto =
-    detailMatch?.params.postId &&
-    photos.find((photo) => photo.id === +detailMatch?.params.postId);
+  const handleCardClick = (postId) => {
+    navigate(`post/${postId}`);
+  };
 
   const handleOverlayClick = () => {
     navigate("/profile");
   };
 
+  const clickedPhoto =
+    detailMatch?.params.postId &&
+    photos.find((photo) => photo.id === +detailMatch?.params.postId);
+
   return (
     <Layout>
       <Container>
-        <AccountInfoTemplate
-          userInfo={userInfo}
-          totalCount={totalCount}
-          isInstaConnected={isInstaConnected}
+        <KaKaoProfile
+          username={username}
+          email={email}
+          profileImage={profileImage}
         />
-        <MyAlbumTemplate photos={photos} />
+        <InstaProfile isLinked={isLinked} infos={infos} />
+        <UploadButton isLinked={isLinked} />
+        <Album>
+          {photos.map((photo) => {
+            return (
+              <Card
+                key={photo.id}
+                photo={photo}
+                layoutId={"my" + photo.id}
+                onClick={() => handleCardClick(photo.id)}
+              />
+            );
+          })}
+        </Album>
       </Container>
       <AnimatePresence>
         {detailMatch && (
@@ -93,8 +175,8 @@ const MyPage = () => {
               animate={{ opacity: 1 }}
               exit={{ opacity: 0 }}
             />
-            <Detail layoutId={"my" + detailMatch?.params.postId}>
-              <MyDetailPage image={clickedPhoto.imgSrc} />
+            <Detail layoutId={"my" + detailMatch.params.postId}>
+              <MyDetailPage photo={clickedPhoto} />
             </Detail>
           </>
         )}
